Semi-Final Update
I'm still fiddling with the format of the result file, but otherwise this is it! The idea of the result file is that it holds the differences between masterlist and comparison. I'm working out the comparison and trying to see if there is a more efficient way to get the differences. Once I get friendlier with TRAVIS, introduced by Arjun to run the code every day, it should be set!